+++ Source/PreviewView.swift
@@ -17,9 +17,32 @@ LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
 import UIKit
 import AVFoundation
 
+/// A function to specifty the Preview Layer's videoGravity. Indicates how the video is displayed within a player layer’s bounds rect.
+@objc public enum SwiftyCamVideoGravity: Int {
+
+    /**
+     - Specifies that the video should be stretched to fill the layer’s bounds
+     - Corrsponds to `AVLayerVideoGravityResize`
+    */
+    case resize
+    /**
+     - Specifies that the player should preserve the video’s aspect ratio and fit the video within the layer’s bounds.
+     - Corresponds to `AVLayerVideoGravityResizeAspect`
+     */
+    case resizeAspect
+    /**
+     - Specifies that the player should preserve the video’s aspect ratio and fill the layer’s bounds.
+     - Correponds to `AVLayerVideoGravityResizeAspectFill`
+    */
+    case resizeAspectFill
+}
+
 @objc open class PreviewView: UIView {
     
-    override init(frame: CGRect) {
+    private var gravity: SwiftyCamVideoGravity = .resizeAspect
+    
+    init(frame: CGRect, videoGravity: SwiftyCamVideoGravity) {
+        gravity = videoGravity
         super.init(frame: frame)
         self.backgroundColor = UIColor.black
     }
@@ -29,7 +52,16 @@ import AVFoundation
     }
     
 	public var videoPreviewLayer: AVCaptureVideoPreviewLayer {
-		return layer as! AVCaptureVideoPreviewLayer
+        let previewlayer = layer as! AVCaptureVideoPreviewLayer
+        switch gravity {
+        case .resize:
+            previewlayer.videoGravity = AVLayerVideoGravity.resize
+        case .resizeAspect:
+            previewlayer.videoGravity = AVLayerVideoGravity.resizeAspect
+        case .resizeAspectFill:
+            previewlayer.videoGravity = AVLayerVideoGravity.resizeAspectFill
+        }
+		return previewlayer
 	}
 	
 	public var session: AVCaptureSession? {