+++ Source/SwiftyCamViewController.swift
@@ -91,7 +91,7 @@ import AVFoundation
 
 	/// Public Camera Delegate for the Custom View Controller Subclass
 
-	public var cameraDelegate: SwiftyCamViewControllerDelegate?
+	public weak var cameraDelegate: SwiftyCamViewControllerDelegate?
 
 	/// Maxiumum video duration if SwiftyCamButton is used
 
@@ -115,7 +115,7 @@ import AVFoundation
 	/// Sets the maximum zoom scale allowed during gestures gesture
 
 	public var maxZoomScale				         = CGFloat.greatestFiniteMagnitude
-  
+
 	/// Sets whether Tap to Focus and Tap to Adjust Exposure is enabled for the capture session
 
 	public var tapToFocus                        = true
@@ -133,13 +133,13 @@ import AVFoundation
 	/// Sets whether a double tap to switch cameras is supported
 
 	public var doubleTapCameraSwitch            = true
-    
+
     /// Sets whether swipe vertically to zoom is supported
-    
+
     public var swipeToZoom                     = true
-    
+
     /// Sets whether swipe vertically gestures should be inverted
-    
+
     public var swipeToZoomInverted             = false
 
 	/// Set default launch camera
@@ -148,7 +148,35 @@ import AVFoundation
 
 	/// Sets wether the taken photo or video should be oriented according to the device orientation
 
-	public var shouldUseDeviceOrientation      = false
+    public var shouldUseDeviceOrientation      = false {
+        didSet {
+            orientation.shouldUseDeviceOrientation = shouldUseDeviceOrientation
+        }
+    }
+
+    /// Sets whether or not View Controller supports auto rotation
+
+    public var allowAutoRotate                = false
+
+    /// Specifies the [videoGravity](https://developer.apple.com/reference/avfoundation/avcapturevideopreviewlayer/1386708-videogravity) for the preview layer.
+    public var videoGravity                   : SwiftyCamVideoGravity = .resizeAspect
+
+    /// Sets whether or not video recordings will record audio
+    /// Setting to true will prompt user for access to microphone on View Controller launch.
+    public var audioEnabled                   = true
+
+    /// Sets whether or not app should display prompt to app settings if audio/video permission is denied
+    /// If set to false, delegate function will be called to handle exception
+    public var shouldPrompToAppSettings       = true
+
+    /// Video will be recorded to this folder
+    public var outputFolder: String           = NSTemporaryDirectory()
+    
+    /// Public access to Pinch Gesture
+    fileprivate(set) public var pinchGesture  : UIPinchGestureRecognizer!
+
+    /// Public access to Pan Gesture
+    fileprivate(set) public var panGesture    : UIPanGestureRecognizer!
 
 
 	// MARK: Public Get-only Variable Declarations
@@ -169,7 +197,7 @@ import AVFoundation
 
 	/// Current Capture Session
 
-	public let session                      = AVCaptureSession()
+	public let session                           = AVCaptureSession()
 
 	/// Serial queue used for setting up session
 
@@ -220,19 +248,23 @@ import AVFoundation
 	/// UIView for front facing flash
 
 	public var flashView                    : UIView?
-    
-  /// Pan Translation
-    
-  public var previousPanTranslation       : CGFloat = 0.0
+
+    /// Pan Translation
+
+    public var previousPanTranslation       : CGFloat = 0.0
 
 	/// Last changed orientation
 
-	public var deviceOrientation            : UIDeviceOrientation?
+    public var orientation                  : Orientation = Orientation()
+
+    /// Boolean to store when View Controller is notified session is running
+
+    public var sessionRunning               = false
 
 	/// Disable view autorotation for forced portrait recorindg
 
 	override open var shouldAutorotate: Bool {
-		return false
+		return allowAutoRotate
 	}
 
 	// MARK: ViewDidLoad
@@ -241,18 +273,20 @@ import AVFoundation
 
 	override open func viewDidLoad() {
 		super.viewDidLoad()
-		previewLayer = PreviewView(frame: CGRect(x: 0.0, y: 0.0, width: view.bounds.width, height: view.bounds.height))
+        previewLayer = PreviewView(frame: view.frame, videoGravity: videoGravity)
+        previewLayer.center = view.center
+        view.addSubview(previewLayer)
+        view.sendSubviewToBack(previewLayer)
 
 		// Add Gesture Recognizers
 
-		addGestureRecognizersTo(view: previewLayer)
+        addGestureRecognizers()
 
-		self.view.addSubview(previewLayer)
 		previewLayer.session = session
 
 		// Test authorization status for Camera and Micophone
 
-		switch AVCaptureDevice.authorizationStatus(forMediaType: AVMediaTypeVideo){
+		switch AVCaptureDevice.authorizationStatus(for: AVMediaType.video) {
 		case .authorized:
 
 			// already authorized
@@ -261,7 +295,7 @@ import AVFoundation
 
 			// not yet determined
 			sessionQueue.suspend()
-			AVCaptureDevice.requestAccess(forMediaType: AVMediaTypeVideo, completionHandler: { [unowned self] granted in
+			AVCaptureDevice.requestAccess(for: AVMediaType.video, completionHandler: { [unowned self] granted in
 				if !granted {
 					self.setupResult = .notAuthorized
 				}
@@ -278,27 +312,74 @@ import AVFoundation
 	}
 
     // MARK: ViewDidLayoutSubviews
-    
+
     /// ViewDidLayoutSubviews() Implementation
-    
-    
-    override open func viewDidLayoutSubviews() {
-        previewLayer.frame = CGRect(x: 0.0, y: 0.0, width: view.bounds.width, height: view.bounds.height)
-        super.viewDidLayoutSubviews()
+    private func updatePreviewLayer(layer: AVCaptureConnection, orientation: AVCaptureVideoOrientation) {
+
+        layer.videoOrientation = orientation
+
+        previewLayer.frame = self.view.bounds
+
     }
-    
+
+    override open func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+
+        if let connection =  self.previewLayer?.videoPreviewLayer.connection  {
+
+            let currentDevice: UIDevice = UIDevice.current
+
+            let orientation: UIDeviceOrientation = currentDevice.orientation
+
+            let previewLayerConnection : AVCaptureConnection = connection
+
+            if previewLayerConnection.isVideoOrientationSupported {
+
+                switch (orientation) {
+                case .portrait: updatePreviewLayer(layer: previewLayerConnection, orientation: .portrait)
+
+                    break
+
+                case .landscapeRight: updatePreviewLayer(layer: previewLayerConnection, orientation: .landscapeLeft)
+
+                    break
+
+                case .landscapeLeft: updatePreviewLayer(layer: previewLayerConnection, orientation: .landscapeRight)
+
+                    break
+
+                case .portraitUpsideDown: updatePreviewLayer(layer: previewLayerConnection, orientation: .portraitUpsideDown)
+
+                    break
+
+                default: updatePreviewLayer(layer: previewLayerConnection, orientation: .portrait)
+
+                    break
+                }
+            }
+        }
+    }
+
+    // MARK: ViewWillAppear
+
+    /// ViewWillAppear(_ animated:) Implementation
+
+    open override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        NotificationCenter.default.addObserver(self, selector: #selector(captureSessionDidStartRunning), name: .AVCaptureSessionDidStartRunning, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(captureSessionDidStopRunning),  name: .AVCaptureSessionDidStopRunning,  object: nil)
+    }
+
 	// MARK: ViewDidAppear
 
 	/// ViewDidAppear(_ animated:) Implementation
-
-
 	override open func viewDidAppear(_ animated: Bool) {
 		super.viewDidAppear(animated)
 
 		// Subscribe to device rotation notifications
 
 		if shouldUseDeviceOrientation {
-			subscribeToDeviceOrientationChangeNotifications()
+			orientation.start()
 		}
 
 		// Set background audio preference
@@ -311,23 +392,23 @@ import AVFoundation
 				// Begin Session
 				self.session.startRunning()
 				self.isSessionRunning = self.session.isRunning
-                
+
                 // Preview layer video orientation can be set only after the connection is created
                 DispatchQueue.main.async {
-                    self.previewLayer.videoPreviewLayer.connection?.videoOrientation = self.getPreviewLayerOrientation()
+                    self.previewLayer.videoPreviewLayer.connection?.videoOrientation = self.orientation.getPreviewLayerOrientation()
                 }
-                
+
 			case .notAuthorized:
-				// Prompt to App Settings
-				self.promptToAppSettings()
+                if self.shouldPrompToAppSettings == true {
+                    self.promptToAppSettings()
+                } else {
+                    self.cameraDelegate?.swiftyCamNotAuthorized(self)
+                }
 			case .configurationFailed:
 				// Unknown Error
-				DispatchQueue.main.async(execute: { [unowned self] in
-					let message = NSLocalizedString("Unable to capture media", comment: "Alert message when something goes wrong during capture session configuration")
-					let alertController = UIAlertController(title: "AVCam", message: message, preferredStyle: .alert)
-					alertController.addAction(UIAlertAction(title: NSLocalizedString("OK", comment: "Alert OK button"), style: .cancel, handler: nil))
-					self.present(alertController, animated: true, completion: nil)
-				})
+                DispatchQueue.main.async {
+                    self.cameraDelegate?.swiftyCamDidFailToConfigure(self)
+                }
 			}
 		}
 	}
@@ -340,6 +421,9 @@ import AVFoundation
 	override open func viewDidDisappear(_ animated: Bool) {
 		super.viewDidDisappear(animated)
 
+        NotificationCenter.default.removeObserver(self)
+        sessionRunning = false
+
 		// If session is running, stop the session
 		if self.isSessionRunning == true {
 			self.session.stopRunning()
@@ -351,7 +435,7 @@ import AVFoundation
 
 		// Unsubscribe from device rotation notifications
 		if shouldUseDeviceOrientation {
-			unsubscribeFromDeviceOrientationChangeNotifications()
+			orientation.stop()
 		}
 	}
 
@@ -370,6 +454,7 @@ import AVFoundation
 		guard let device = videoDevice else {
 			return
 		}
+
 
 		if device.hasFlash == true && flashEnabled == true /* TODO: Add Support for Retina Flash and add front flash */ {
 			changeFlashSettings(device: device, mode: .on)
@@ -410,6 +495,11 @@ import AVFoundation
 	*/
 
 	public func startVideoRecording() {
+
+        guard sessionRunning == true else {
+            print("[SwiftyCam]: Cannot start video recoding. Capture session is not running")
+            return
+        }
 		guard let movieFileOutput = self.movieFileOutput else {
 			return
 		}
@@ -425,6 +515,9 @@ import AVFoundation
 			previewLayer.addSubview(flashView!)
 		}
 
+        //Must be fetched before on main thread
+        let previewOrientation = previewLayer.videoPreviewLayer.connection!.videoOrientation
+
 		sessionQueue.async { [unowned self] in
 			if !movieFileOutput.isRecording {
 				if UIDevice.current.isMultitaskingSupported {
@@ -432,7 +525,7 @@ import AVFoundation
 				}
 
 				// Update the orientation on the movie file output video connection before starting recording.
-				let movieFileOutputConnection = self.movieFileOutput?.connection(withMediaType: AVMediaTypeVideo)
+				let movieFileOutputConnection = self.movieFileOutput?.connection(with: AVMediaType.video)
 
 
 				//flip video output if front facing camera is selected
@@ -440,12 +533,12 @@ import AVFoundation
 					movieFileOutputConnection?.isVideoMirrored = true
 				}
 
-				movieFileOutputConnection?.videoOrientation = self.getVideoOrientation()
+				movieFileOutputConnection?.videoOrientation = self.orientation.getVideoOrientation() ?? previewOrientation
 
 				// Start recording to a temporary file.
 				let outputFileName = UUID().uuidString
-				let outputFilePath = (NSTemporaryDirectory() as NSString).appendingPathComponent((outputFileName as NSString).appendingPathExtension("mov")!)
-				movieFileOutput.startRecording(toOutputFileURL: URL(fileURLWithPath: outputFilePath), recordingDelegate: self)
+				let outputFilePath = (self.outputFolder as NSString).appendingPathComponent((outputFileName as NSString).appendingPathExtension("mov")!)
+				movieFileOutput.startRecording(to: URL(fileURLWithPath: outputFilePath), recordingDelegate: self)
 				self.isVideoRecording = true
 				DispatchQueue.main.async {
 					self.cameraDelegate?.swiftyCam(self, didBeginRecordingVideo: self.currentCamera)
@@ -468,7 +561,7 @@ import AVFoundation
 	*/
 
 	public func stopVideoRecording() {
-		if self.movieFileOutput?.isRecording == true {
+		if self.isVideoRecording == true {
 			self.isVideoRecording = false
 			movieFileOutput!.stopRecording()
 			disableFlash()
@@ -501,11 +594,11 @@ import AVFoundation
 			print("[SwiftyCam]: Switching between cameras while recording video is not supported")
 			return
 		}
-        
+
         guard session.isRunning == true else {
             return
         }
-        
+
 		switch currentCamera {
 		case .front:
 			currentCamera = .rear
@@ -520,7 +613,7 @@ import AVFoundation
 			// remove and re-add inputs and outputs
 
 			for input in self.session.inputs {
-				self.session.removeInput(input as! AVCaptureInput)
+				self.session.removeInput(input )
 			}
 
 			self.addInputs()
@@ -580,14 +673,13 @@ import AVFoundation
 	/// Configure image quality preset
 
 	fileprivate func configureVideoPreset() {
-
 		if currentCamera == .front {
-			session.sessionPreset = videoInputPresetFromVideoQuality(quality: .high)
+			session.sessionPreset = AVCaptureSession.Preset(rawValue: videoInputPresetFromVideoQuality(quality: .high))
 		} else {
-			if session.canSetSessionPreset(videoInputPresetFromVideoQuality(quality: videoQuality)) {
-				session.sessionPreset = videoInputPresetFromVideoQuality(quality: videoQuality)
+			if session.canSetSessionPreset(AVCaptureSession.Preset(rawValue: videoInputPresetFromVideoQuality(quality: videoQuality))) {
+				session.sessionPreset = AVCaptureSession.Preset(rawValue: videoInputPresetFromVideoQuality(quality: videoQuality))
 			} else {
-				session.sessionPreset = videoInputPresetFromVideoQuality(quality: .high)
+				session.sessionPreset = AVCaptureSession.Preset(rawValue: videoInputPresetFromVideoQuality(quality: .high))
 			}
 		}
 	}
@@ -597,9 +689,9 @@ import AVFoundation
 	fileprivate func addVideoInput() {
 		switch currentCamera {
 		case .front:
-			videoDevice = SwiftyCamViewController.deviceWithMediaType(AVMediaTypeVideo, preferringPosition: .front)
+			videoDevice = SwiftyCamViewController.deviceWithMediaType(AVMediaType.video.rawValue, preferringPosition: .front)
 		case .rear:
-			videoDevice = SwiftyCamViewController.deviceWithMediaType(AVMediaTypeVideo, preferringPosition: .back)
+			videoDevice = SwiftyCamViewController.deviceWithMediaType(AVMediaType.video.rawValue, preferringPosition: .back)
 		}
 
 		if let device = videoDevice {
@@ -631,18 +723,20 @@ import AVFoundation
 		}
 
 		do {
-			let videoDeviceInput = try AVCaptureDeviceInput(device: videoDevice)
-
-			if session.canAddInput(videoDeviceInput) {
-				session.addInput(videoDeviceInput)
-				self.videoDeviceInput = videoDeviceInput
-			} else {
-				print("[SwiftyCam]: Could not add video device input to the session")
-				print(session.canSetSessionPreset(videoInputPresetFromVideoQuality(quality: videoQuality)))
-				setupResult = .configurationFailed
-				session.commitConfiguration()
-				return
-			}
+            if let videoDevice = videoDevice {
+                let videoDeviceInput = try AVCaptureDeviceInput(device: videoDevice)
+                if session.canAddInput(videoDeviceInput) {
+                    session.addInput(videoDeviceInput)
+                    self.videoDeviceInput = videoDeviceInput
+                } else {
+                    print("[SwiftyCam]: Could not add video device input to the session")
+                    print(session.canSetSessionPreset(AVCaptureSession.Preset(rawValue: videoInputPresetFromVideoQuality(quality: videoQuality))))
+                    setupResult = .configurationFailed
+                    session.commitConfiguration()
+                    return
+                }
+            }
+			
 		} catch {
 			print("[SwiftyCam]: Could not create video device input: \(error)")
 			setupResult = .configurationFailed
@@ -653,18 +747,23 @@ import AVFoundation
 	/// Add Audio Inputs
 
 	fileprivate func addAudioInput() {
+        guard audioEnabled == true else {
+            return
+        }
 		do {
-			let audioDevice = AVCaptureDevice.defaultDevice(withMediaType: AVMediaTypeAudio)
-			let audioDeviceInput = try AVCaptureDeviceInput(device: audioDevice)
-
-			if session.canAddInput(audioDeviceInput) {
-				session.addInput(audioDeviceInput)
-			}
-			else {
-				print("[SwiftyCam]: Could not add audio device input to the session")
-			}
-		}
-		catch {
+            if let audioDevice = AVCaptureDevice.default(for: AVMediaType.audio){
+                let audioDeviceInput = try AVCaptureDeviceInput(device: audioDevice)
+                if session.canAddInput(audioDeviceInput) {
+                    session.addInput(audioDeviceInput)
+                } else {
+                    print("[SwiftyCam]: Could not add audio device input to the session")
+                }
+                
+            } else {
+                print("[SwiftyCam]: Could not find an audio device")
+            }
+            
+		} catch {
 			print("[SwiftyCam]: Could not create audio device input: \(error)")
 		}
 	}
@@ -676,7 +775,7 @@ import AVFoundation
 
 		if self.session.canAddOutput(movieFileOutput) {
 			self.session.addOutput(movieFileOutput)
-			if let connection = movieFileOutput.connection(withMediaType: AVMediaTypeVideo) {
+			if let connection = movieFileOutput.connection(with: AVMediaType.video) {
 				if connection.isVideoStabilizationSupported {
 					connection.preferredVideoStabilizationMode = .auto
 				}
@@ -699,67 +798,67 @@ import AVFoundation
 
 	/// Orientation management
 
-	@objc public func subscribeToDeviceOrientationChangeNotifications() {
-		self.deviceOrientation = UIDevice.current.orientation
-		NotificationCenter.default.addObserver(self, selector: #selector(deviceDidRotate), name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
-	}
+	// @objc public func subscribeToDeviceOrientationChangeNotifications() {
+	// 	self.deviceOrientation = UIDevice.current.orientation
+	// 	NotificationCenter.default.addObserver(self, selector: #selector(deviceDidRotate), name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
+	// }
 
-	@objc public func unsubscribeFromDeviceOrientationChangeNotifications() {
-		NotificationCenter.default.removeObserver(self, name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
-		self.deviceOrientation = nil
-	}
+	// @objc public func unsubscribeFromDeviceOrientationChangeNotifications() {
+	// 	NotificationCenter.default.removeObserver(self, name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
+	// 	self.deviceOrientation = nil
+	// }
 
-	@objc public func deviceDidRotate() {
-		if !UIDevice.current.orientation.isFlat {
-			self.deviceOrientation = UIDevice.current.orientation
-		}
-	}
+	// @objc public func deviceDidRotate() {
+	// 	if !UIDevice.current.orientation.isFlat {
+	// 		self.deviceOrientation = UIDevice.current.orientation
+	// 	}
+	// }
     
-    @objc public func getPreviewLayerOrientation() -> AVCaptureVideoOrientation {
-        // Depends on layout orientation, not device orientation
-        switch UIApplication.shared.statusBarOrientation {
-        case .portrait, .unknown:
-            return AVCaptureVideoOrientation.portrait
-        case .landscapeLeft:
-            return AVCaptureVideoOrientation.landscapeLeft
-        case .landscapeRight:
-            return AVCaptureVideoOrientation.landscapeRight
-        case .portraitUpsideDown:
-            return AVCaptureVideoOrientation.portraitUpsideDown
-        }
-    }
+  //   @objc public func getPreviewLayerOrientation() -> AVCaptureVideoOrientation {
+  //       // Depends on layout orientation, not device orientation
+  //       switch UIApplication.shared.statusBarOrientation {
+  //       case .portrait, .unknown:
+  //           return AVCaptureVideoOrientation.portrait
+  //       case .landscapeLeft:
+  //           return AVCaptureVideoOrientation.landscapeLeft
+  //       case .landscapeRight:
+  //           return AVCaptureVideoOrientation.landscapeRight
+  //       case .portraitUpsideDown:
+  //           return AVCaptureVideoOrientation.portraitUpsideDown
+  //       }
+  //   }
 
-	@objc public func getVideoOrientation() -> AVCaptureVideoOrientation {
-		guard shouldUseDeviceOrientation, let deviceOrientation = self.deviceOrientation else { return previewLayer!.videoPreviewLayer.connection.videoOrientation }
+	// @objc public func getVideoOrientation() -> AVCaptureVideoOrientation {
+	// 	guard shouldUseDeviceOrientation, let deviceOrientation = self.deviceOrientation else { return previewLayer!.videoPreviewLayer.connection.videoOrientation }
 
-		switch deviceOrientation {
-		case .landscapeLeft:
-			// keep the same if using front camera
-			return self.currentCamera == .rear ? .landscapeRight : .landscapeLeft;
-		case .landscapeRight:
-			// keep the same if using front camera
-			return self.currentCamera == .rear ? .landscapeLeft : .landscapeRight;
-		case .portraitUpsideDown:
-			return .portraitUpsideDown
-		default:
-			return .portrait
-		}
-	}
+	// 	switch deviceOrientation {
+	// 	case .landscapeLeft:
+	// 		// keep the same if using front camera
+	// 		return self.currentCamera == .rear ? .landscapeRight : .landscapeLeft;
+	// 	case .landscapeRight:
+	// 		// keep the same if using front camera
+	// 		return self.currentCamera == .rear ? .landscapeLeft : .landscapeRight;
+	// 	case .portraitUpsideDown:
+	// 		return .portraitUpsideDown
+	// 	default:
+	// 		return .portrait
+	// 	}
+	// }
 
-	@objc public func getImageOrientation(forCamera: CameraSelection) -> UIImageOrientation {
-		guard shouldUseDeviceOrientation, let deviceOrientation = self.deviceOrientation else { return forCamera == .rear ? .right : .leftMirrored }
+	// @objc public func getImageOrientation(forCamera: CameraSelection) -> UIImageOrientation {
+	// 	guard shouldUseDeviceOrientation, let deviceOrientation = self.deviceOrientation else { return forCamera == .rear ? .right : .leftMirrored }
 
-		switch deviceOrientation {
-		case .landscapeLeft:
-			return forCamera == .rear ? .up : .downMirrored
-		case .landscapeRight:
-			return forCamera == .rear ? .down : .upMirrored
-		case .portraitUpsideDown:
-			return forCamera == .rear ? .left : .rightMirrored
-		default:
-			return forCamera == .rear ? .right : .leftMirrored
-		}
-	}
+	// 	switch deviceOrientation {
+	// 	case .landscapeLeft:
+	// 		return forCamera == .rear ? .up : .downMirrored
+	// 	case .landscapeRight:
+	// 		return forCamera == .rear ? .down : .upMirrored
+	// 	case .portraitUpsideDown:
+	// 		return forCamera == .rear ? .left : .rightMirrored
+	// 	default:
+	// 		return forCamera == .rear ? .right : .leftMirrored
+	// 	}
+	// }
 
 	/**
 	Returns a UIImage from Image Data.
@@ -776,17 +875,23 @@ import AVFoundation
 		// Set proper orientation for photo
 		// If camera is currently set to front camera, flip image
 
-		let image = UIImage(cgImage: cgImageRef!, scale: 1.0, orientation: self.getImageOrientation(forCamera: self.currentCamera))
+		let image = UIImage(cgImage: cgImageRef!, scale: 1.0, orientation: self.orientation.getImageOrientation(forCamera: self.currentCamera))
 
 		return image
 	}
 
 	@objc public func capturePhotoAsyncronously(completionHandler: @escaping(Bool) -> ()) {
-		if let videoConnection = photoFileOutput?.connection(withMediaType: AVMediaTypeVideo) {
+
+        guard sessionRunning == true else {
+            print("[SwiftyCam]: Cannot take photo. Capture session is not running")
+            return
+        }
+
+		if let videoConnection = photoFileOutput?.connection(with: AVMediaType.video) {
 
 			photoFileOutput?.captureStillImageAsynchronously(from: videoConnection, completionHandler: {(sampleBuffer, error) in
 				if (sampleBuffer != nil) {
-					let imageData = AVCaptureStillImageOutput.jpegStillImageNSDataRepresentation(sampleBuffer)
+					let imageData = AVCaptureStillImageOutput.jpegStillImageNSDataRepresentation(sampleBuffer!)
 					let image = self.processPhoto(imageData!)
 
 					// Call delegate and return new image
@@ -814,9 +919,9 @@ import AVFoundation
 			alertController.addAction(UIAlertAction(title: NSLocalizedString("OK", comment: "Alert OK button"), style: .cancel, handler: nil))
 			alertController.addAction(UIAlertAction(title: NSLocalizedString("Settings", comment: "Alert button to open Settings"), style: .default, handler: { action in
 				if #available(iOS 10.0, *) {
-					UIApplication.shared.openURL(URL(string: UIApplicationOpenSettingsURLString)!)
+					UIApplication.shared.openURL(URL(string: UIApplication.openSettingsURLString)!)
 				} else {
-					if let appSettings = URL(string: UIApplicationOpenSettingsURLString) {
+					if let appSettings = URL(string: UIApplication.openSettingsURLString) {
 						UIApplication.shared.openURL(appSettings)
 					}
 				}
@@ -835,38 +940,54 @@ import AVFoundation
 
 	@objc public func videoInputPresetFromVideoQuality(quality: VideoQuality) -> String {
 		switch quality {
-		case .high: return AVCaptureSessionPresetHigh
-		case .medium: return AVCaptureSessionPresetMedium
-		case .low: return AVCaptureSessionPresetLow
-		case .resolution352x288: return AVCaptureSessionPreset352x288
-		case .resolution640x480: return AVCaptureSessionPreset640x480
-		case .resolution1280x720: return AVCaptureSessionPreset1280x720
-		case .resolution1920x1080: return AVCaptureSessionPreset1920x1080
-		case .iframe960x540: return AVCaptureSessionPresetiFrame960x540
-		case .iframe1280x720: return AVCaptureSessionPresetiFrame1280x720
+		case .high: return AVCaptureSession.Preset.high.rawValue
+		case .medium: return AVCaptureSession.Preset.medium.rawValue
+		case .low: return AVCaptureSession.Preset.low.rawValue
+		case .resolution352x288: return AVCaptureSession.Preset.cif352x288.rawValue
+		case .resolution640x480: return AVCaptureSession.Preset.vga640x480.rawValue
+		case .resolution1280x720: return AVCaptureSession.Preset.hd1280x720.rawValue
+		case .resolution1920x1080: return AVCaptureSession.Preset.hd1920x1080.rawValue
+		case .iframe960x540: return AVCaptureSession.Preset.iFrame960x540.rawValue
+		case .iframe1280x720: return AVCaptureSession.Preset.iFrame1280x720.rawValue
 		case .resolution3840x2160:
 			if #available(iOS 9.0, *) {
-				return AVCaptureSessionPreset3840x2160
+				return AVCaptureSession.Preset.hd4K3840x2160.rawValue
 			}
 			else {
 				print("[SwiftyCam]: Resolution 3840x2160 not supported")
-				return AVCaptureSessionPresetHigh
+				return AVCaptureSession.Preset.high.rawValue
 			}
 		}
 	}
 
 	/// Get Devices
 
-	@objc public class func deviceWithMediaType(_ mediaType: String, preferringPosition position: AVCaptureDevicePosition) -> AVCaptureDevice? {
-		if let devices = AVCaptureDevice.devices(withMediaType: mediaType) as? [AVCaptureDevice] {
-			return devices.filter({ $0.position == position }).first
+	@objc public class func deviceWithMediaType(_ mediaType: String, preferringPosition position: AVCaptureDevice.Position) -> AVCaptureDevice? {
+		if #available(iOS 10.0, *) {
+				let avDevice = AVCaptureDevice.default(AVCaptureDevice.DeviceType.builtInWideAngleCamera, for: AVMediaType(rawValue: mediaType), position: position)
+				return avDevice
+		} else {
+				// Fallback on earlier versions
+				let avDevice = AVCaptureDevice.devices(for: AVMediaType(rawValue: mediaType))
+				var avDeviceNum = 0
+				for device in avDevice {
+						print("deviceWithMediaType Position: \(device.position.rawValue)")
+						if device.position == position {
+								break
+						} else {
+								avDeviceNum += 1
+						}
+				}
+
+				return avDevice[avDeviceNum]
 		}
-		return nil
+
+		//return AVCaptureDevice.devices(for: AVMediaType(rawValue: mediaType), position: position).first
 	}
 
 	/// Enable or disable flash for photo
 
-	@objc public func changeFlashSettings(device: AVCaptureDevice, mode: AVCaptureFlashMode) {
+	@objc public func changeFlashSettings(device: AVCaptureDevice, mode: AVCaptureDevice.FlashMode) {
 		do {
 			try device.lockForConfiguration()
 			device.flashMode = mode
@@ -900,17 +1021,17 @@ import AVFoundation
 			return
 		}
 
-		let device = AVCaptureDevice.defaultDevice(withMediaType: AVMediaTypeVideo)
+		let device = AVCaptureDevice.default(for: AVMediaType.video)
 		// Check if device has a flash
 		if (device?.hasTorch)! {
 			do {
 				try device?.lockForConfiguration()
-				if (device?.torchMode == AVCaptureTorchMode.on) {
-					device?.torchMode = AVCaptureTorchMode.off
+				if (device?.torchMode == AVCaptureDevice.TorchMode.on) {
+					device?.torchMode = AVCaptureDevice.TorchMode.off
 					self.isCameraTorchOn = false
 				} else {
 					do {
-						try device?.setTorchModeOnWithLevel(1.0)
+						try device?.setTorchModeOn(level: 1.0)
 						self.isCameraTorchOn = true
 					} catch {
 						print("[SwiftyCam]: \(error)")
@@ -930,10 +1051,20 @@ import AVFoundation
 			return
 		}
 
-		do{
-			try AVAudioSession.sharedInstance().setCategory(AVAudioSessionCategoryPlayAndRecord,
-			                                                with: [.duckOthers, .defaultToSpeaker])
+        guard audioEnabled == true else {
+            return
+        }
 
+		do{
+            if #available(iOS 10.0, *) {
+                try AVAudioSession.sharedInstance().setCategory(.playAndRecord, mode: .default, options: [.mixWithOthers, .allowBluetooth, .allowAirPlay, .allowBluetoothA2DP])
+            } else {
+                let options: [AVAudioSession.CategoryOptions] = [.mixWithOthers, .allowBluetooth]
+                              let category = AVAudioSession.Category.playAndRecord
+                let selector = NSSelectorFromString("setCategory:withOptions:error:")
+                AVAudioSession.sharedInstance().perform(selector, with: category, with: options)
+            }
+            try AVAudioSession.sharedInstance().setActive(true)
 			session.automaticallyConfiguresApplicationAudioSession = false
 		}
 		catch {
@@ -941,6 +1072,24 @@ import AVFoundation
 
 		}
 	}
+
+    /// Called when Notification Center registers session starts running
+
+    @objc private func captureSessionDidStartRunning() {
+        sessionRunning = true
+        DispatchQueue.main.async {
+            self.cameraDelegate?.swiftyCamSessionDidStartRunning(self)
+        }
+    }
+
+    /// Called when Notification Center registers session stops running
+
+    @objc private func captureSessionDidStopRunning() {
+        sessionRunning = false
+        DispatchQueue.main.async {
+            self.cameraDelegate?.swiftyCamSessionDidStopRunning(self)
+        }
+    }
 }
 
 extension SwiftyCamViewController : SwiftyCamButtonDelegate {
@@ -983,23 +1132,27 @@ extension SwiftyCamViewController : AVCaptureFileOutputRecordingDelegate {
 
 	/// Process newly captured video and write it to temporary directory
 
-	public func capture(_ captureOutput: AVCaptureFileOutput!, didFinishRecordingToOutputFileAt outputFileURL: URL!, fromConnections connections: [Any]!, error: Error!) {
-		if let currentBackgroundRecordingID = backgroundRecordingID {
-			backgroundRecordingID = UIBackgroundTaskInvalid
+    public func fileOutput(_ output: AVCaptureFileOutput, didFinishRecordingTo outputFileURL: URL, from connections: [AVCaptureConnection], error: Error?) {
+        if let currentBackgroundRecordingID = backgroundRecordingID {
+            backgroundRecordingID = UIBackgroundTaskIdentifier.invalid
 
-			if currentBackgroundRecordingID != UIBackgroundTaskInvalid {
-				UIApplication.shared.endBackgroundTask(currentBackgroundRecordingID)
-			}
-		}
-		if error != nil {
-			print("[SwiftyCam]: Movie file finishing error: \(error)")
-		} else {
-			//Call delegate function with the URL of the outputfile
-			DispatchQueue.main.async {
-				self.cameraDelegate?.swiftyCam(self, didFinishProcessVideoAt: outputFileURL)
-			}
-		}
-	}
+            if currentBackgroundRecordingID != UIBackgroundTaskIdentifier.invalid {
+                UIApplication.shared.endBackgroundTask(currentBackgroundRecordingID)
+            }
+        }
+
+        if let currentError = error {
+            print("[SwiftyCam]: Movie file finishing error: \(currentError)")
+            DispatchQueue.main.async {
+                self.cameraDelegate?.swiftyCam(self, didFailToRecordVideo: currentError)
+            }
+        } else {
+            //Call delegate function with the URL of the outputfile
+            DispatchQueue.main.async {
+                self.cameraDelegate?.swiftyCam(self, didFinishProcessVideoAt: outputFileURL)
+            }
+        }
+    }
 }
 
 // Mark: UIGestureRecognizer Declarations
@@ -1010,11 +1163,11 @@ extension SwiftyCamViewController {
 
 	@objc fileprivate func zoomGesture(pinch: UIPinchGestureRecognizer) {
 		guard pinchToZoom == true && self.currentCamera == .rear else {
-			//ignore pinch 
+			//ignore pinch
 			return
 		}
 		do {
-			let captureDevice = AVCaptureDevice.devices().first as? AVCaptureDevice
+			let captureDevice = AVCaptureDevice.devices().first
 			try captureDevice?.lockForConfiguration()
 
 			zoomScale = min(maxZoomScale, max(1.0, min(beginZoomScale * pinch.scale,  captureDevice!.activeFormat.videoMaxZoomFactor)))
@@ -1056,7 +1209,7 @@ extension SwiftyCamViewController {
 					device.focusMode = .autoFocus
 				}
 				device.exposurePointOfInterest = focusPoint
-				device.exposureMode = AVCaptureExposureMode.continuousAutoExposure
+				device.exposureMode = AVCaptureDevice.ExposureMode.continuousAutoExposure
 				device.unlockForConfiguration()
 				//Call delegate function and pass in the location of the touch
 
@@ -1078,42 +1231,42 @@ extension SwiftyCamViewController {
 		}
 		switchCamera()
 	}
-    
+
     @objc private func panGesture(pan: UIPanGestureRecognizer) {
-        
+
         guard swipeToZoom == true && self.currentCamera == .rear else {
             //ignore pan
             return
         }
         let currentTranslation    = pan.translation(in: view).y
         let translationDifference = currentTranslation - previousPanTranslation
-        
+
         do {
-            let captureDevice = AVCaptureDevice.devices().first as? AVCaptureDevice
+            let captureDevice = AVCaptureDevice.devices().first
             try captureDevice?.lockForConfiguration()
-            
+
             let currentZoom = captureDevice?.videoZoomFactor ?? 0.0
-            
+
             if swipeToZoomInverted == true {
                 zoomScale = min(maxZoomScale, max(1.0, min(currentZoom - (translationDifference / 75),  captureDevice!.activeFormat.videoMaxZoomFactor)))
             } else {
                 zoomScale = min(maxZoomScale, max(1.0, min(currentZoom + (translationDifference / 75),  captureDevice!.activeFormat.videoMaxZoomFactor)))
 
             }
-            
+
             captureDevice?.videoZoomFactor = zoomScale
-            
+
             // Call Delegate function with current zoom scale
             DispatchQueue.main.async {
                 self.cameraDelegate?.swiftyCam(self, didChangeZoomLevel: self.zoomScale)
             }
-            
+
             captureDevice?.unlockForConfiguration()
-            
+
         } catch {
             print("[SwiftyCam]: Error locking configuration")
         }
-        
+
         if pan.state == .ended || pan.state == .failed || pan.state == .cancelled {
             previousPanTranslation = 0.0
         } else {
@@ -1128,24 +1281,24 @@ extension SwiftyCamViewController {
 
 	*/
 
-	fileprivate func addGestureRecognizersTo(view: UIView) {
-		let pinchGesture = UIPinchGestureRecognizer(target: self, action: #selector(zoomGesture(pinch:)))
+	fileprivate func addGestureRecognizers() {
+		pinchGesture = UIPinchGestureRecognizer(target: self, action: #selector(zoomGesture(pinch:)))
 		pinchGesture.delegate = self
-		view.addGestureRecognizer(pinchGesture)
+		previewLayer.addGestureRecognizer(pinchGesture)
 
 		let singleTapGesture = UITapGestureRecognizer(target: self, action: #selector(singleTapGesture(tap:)))
 		singleTapGesture.numberOfTapsRequired = 1
 		singleTapGesture.delegate = self
-		view.addGestureRecognizer(singleTapGesture)
+		previewLayer.addGestureRecognizer(singleTapGesture)
 
 		let doubleTapGesture = UITapGestureRecognizer(target: self, action: #selector(doubleTapGesture(tap:)))
 		doubleTapGesture.numberOfTapsRequired = 2
 		doubleTapGesture.delegate = self
-		view.addGestureRecognizer(doubleTapGesture)
-        
-        let panGesture = UIPanGestureRecognizer(target: self, action: #selector(panGesture(pan:)))
+		previewLayer.addGestureRecognizer(doubleTapGesture)
+
+        panGesture = UIPanGestureRecognizer(target: self, action: #selector(panGesture(pan:)))
         panGesture.delegate = self
-        view.addGestureRecognizer(panGesture)
+        previewLayer.addGestureRecognizer(panGesture)
 	}
 }
 
@@ -1163,4 +1316,3 @@ extension SwiftyCamViewController : UIGestureRecognizerDelegate {
 		return true
 	}
 }
-